requirements-txt.txt
torch==2.2.2
torchaudio==2.2.2
transformers==4.40.1
datasets==2.18.0
evaluate==0.4.2
jiwer==3.0.3
librosa==0.10.2
numpy==1.26.4
pandas==2.2.2
pyyaml==6.0.1
tensorboard==2.16.2
fsspec==2024.3.1
accelerate==0.28.0
omegaconf==2.3.0
soundfile==0.12.1
